+++ CHANGELOG.md
@@ -11,11 +11,13 @@ I figured that this would be nice to have, as to easily take a look at everythin
 
 ### New Stuff!
 - Added `DamagePlayerOnImpact` for `AddRigidbody2D` buildingEntites
-- `setbodyfield` and `setlimbfield` can now change fields on attached CUCoreLib statuses with `StatusType.Field`.
+- `setbodyfield` and `setlimbfield` can now change fields on attached CUCoreLib statuses with `StatusType.Field`. (I.e. ToxicStatuses.Toxicity)
 
 ### Changes
 
 ### Fixes
+- ""Fixed"" large amounts of data being truncated when sending over multiplayer requests
+- Explicit MP support for buildingEntites, tiles. Note this might break for people using v1.0.4/1.0.5 <-> v1.0.6 in terms of crossplay
 - Fixed `AddRigidbody2D` (t'was a layer issue)
 
 
@@ -51,6 +53,7 @@ I figured that this would be nice to have, as to easily take a look at everythin
 - Spritesheet support (very basic)
 
 ### Changes
+- Structure markers now resolve registered CUCoreLib tile and building-entity IDs interchangeably, using marker type to choose priority
 - QoL settings menu compatibility
 - Fixed ScaleWithCondition scaling towards 0.1f. Technically this is basegame, but a bit confusing to most
 - Tiles now can use alphanumeric IDs `TileRegistry.Register("auric", ... )` (o7 calamity)